# expression: implement vectorized evaluation for `builtinCastIntAsDecimalSig`