--- a/expression/builtin_cast_vec.go
+++ b/expression/builtin_cast_vec.go
@@ -813,10 +813,11 @@ func (b *builtinCastRealAsDecimalSig) vecEvalDecimal(input *chunk.Chunk, result 
 				return err
 			}
 		}
-		_, err = types.ProduceDecWithSpecifiedTp(&resdecimal[i], b.tp, b.ctx.GetSessionVars().StmtCtx)
+		dec, err := types.ProduceDecWithSpecifiedTp(&resdecimal[i], b.tp, b.ctx.GetSessionVars().StmtCtx)
 		if err != nil {
 			return err
 		}
+		resdecimal[i] = *dec
 	}
 	return nil
 }