--- a/expression/builtin_cast_vec_test.go
+++ b/expression/builtin_cast_vec_test.go
@@ -90,6 +90,7 @@ var vecBuiltinCastCases = map[string][]vecExprBenchCase{
 				&dateStrGener{},
 				&timeStrGener{},
 			}},
+<<<<<<< HEAD
 		{retEvalType: types.ETDatetime, childrenTypes: []types.EvalType{types.ETDuration}},
 		{retEvalType: types.ETDatetime, childrenTypes: []types.EvalType{types.ETDatetime}},
 		{retEvalType: types.ETDatetime, childrenTypes: []types.EvalType{types.ETTimestamp}},
@@ -98,6 +99,7 @@ var vecBuiltinCastCases = map[string][]vecExprBenchCase{
 				&jsonTimeGener{},
 			}},
 		{retEvalType: types.ETDecimal, childrenTypes: []types.EvalType{types.ETDecimal}},
+		{retEvalType: types.ETDecimal, childrenTypes: []types.EvalType{types.ETInt}},
 	},
 }