PCP #12102

## What have you changed?

implement vectorized evaluation for `builtinCastIntAsDecimalSig`

## What is changed and how it works?
```bash
$$ go test -check.f TestVectorizedBuiltinCastFunc
PASS: builtin_cast_vec_test.go:104: testEvaluatorSuite.TestVectorizedBuiltinCastFunc    0.111s
OK: 1 passed
PASS
ok      github.com/pingcap/tidb/expression      0.144s

$$ go test -v -benchmem -bench=BenchmarkVectorizedBuiltinCastFunc -run=BenchmarkVectorizedBuiltinCastFunc -args "builtinCastIntAsDecimalSig"
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinCastFunc/builtinCastIntAsDecimalSig-VecBuiltinFunc-8             45775             25758 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinCastFunc/builtinCastIntAsDecimalSig-NonVecBuiltinFunc-8          15553             76554 ns/op           40176 B/op        837 allocs/op
PASS
ok      github.com/pingcap/tidb/expression      3.453s

```

## Check List

Tests
  - Unit test