@qw4990 2019-12-19
I will investigate this CI problem soon.