@qw4990 2019-12-20
The CI problem will be fixed by https://github.com/pingcap/tidb/pull/14156.